pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apple-calendar-multi-mcp"
version = "0.1.0"
description = "Multi-account Apple Calendar MCP server backed by iCloud CalDAV"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "caldav>=1.3.9",
    "icalendar>=5.0.11",
    "mcp>=1.23.0",
    "PyYAML>=6.0.2",
    "python-dotenv==1.0.1",
    "requests>=2.32.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
apple-calendar-multi-mcp = "apple_calendar_multi_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/apple_calendar_multi_mcp"]
